Moved code to new skinning class.
